requirements.txt
<<<<<<< Updated upstream
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.8.4
wordcloud==1.9.3
=======
apache-airflow==2.9.1
apache-airflow-providers-postgres==5.7.1

psycopg2-binary==2.9.9
SQLAlchemy==2.0.30

pandas==2.2.2
requests==2.32.3

python-dotenv==1.0.0
rich==13.7.1
>>>>>>> Stashed changes
